back/pyproject.toml
[tool.poetry]
name = "tmeit_backend"
version = "0.1.0"
description = "Python backend for the TMEIT website"
authors = ["TraditionsMEsterIT"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^1.1.1"
Flask-SQLAlchemy = "^2.4.0"
PyJWT = "^1.7.1"
requests = "^2.21.0"
SQLAlchemy = "^1.3.12"
marshmallow = "^3.3.0"
flask-marshmallow = "^0.10.1"
marshmallow-sqlalchemy = "^0.21.0"
marshmallow-enum = "^1.5.1"
argon2-cffi = "^19.2.0"
factory-boy = "^2.12.0"

[tool.poetry.dev-dependencies]
coverage = "^4.5.2"
pytest = "^4.2.0"
pytest-cov = "^2.6.1"
pytest-flask = "^0.14.0"
python-coveralls = "^2.9.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
